requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.4
Django==4.2.6
django-cors-headers==4.3.0
django-filter==23.3
django-js-asset==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.6
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.2.0
